fts-postgres/language-detection/requirements.txt
langdetect==1.0.9
psycopg2-binary==2.9.10
six==1.17.0
